+++ module/sheets/actor-sheet.mjs
@@ -32,7 +32,10 @@ export class BitdActorSheet extends ActorSheet
     const actorData = this.actor.toObject(false);
 
     // Encrich editor content
-    context.enrichedDescription = await TextEditor.enrichHTML(this.object.system.description, { async: true })
+    context.enrichedDescription = await TextEditor.enrichHTML(this.object.system.description, {
+      async: true,
+      secrets: this.actor.isOwner
+    })
 
     // Add the actor's data to context.data for easier access, as well as flags.
     context.system = actorData.system;