+++ CHANGELOG.md
@@ -14,4 +14,9 @@ First official version
 - Change crew header grid and add ability to set a profile picture
 - Contacts: choose type, not input. Description > Summarize
 - Margin between rows in links
-- Show secrets for owner and GM
+- Show secrets only for owner and GM
+
+#### Claims
+
+- Claims as item in crew sheet
+- Claims as link in crew type sheet